server/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
arrow==0.10.0
decorator==4.0.11         # via ipython, traitlets
django-cors-headers==2.1.0
django==1.11.2
djangorestframework==3.6.3
ipython-genutils==0.2.0   # via traitlets
ipython==6.1.0
jedi==0.10.2              # via ipython
pexpect==4.2.1            # via ipython
pickleshare==0.7.4        # via ipython
prompt-toolkit==1.0.14    # via ipython
psycopg2==2.7.1
ptyprocess==0.5.1         # via pexpect
pygments==2.2.0           # via ipython
python-dateutil==2.6.0    # via arrow
pytz==2017.2              # via django
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via prompt-toolkit, python-dateutil, traitlets
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
